feat: add 2 versions of a MaixSenseA010ImageQueue
We redefine MaixSenseA010ImageQueue as an interface. Then, we define
- MaixSenseA010ImagePlainQueue, that simply adds and polls
MaixSenseA010Images
- MaixSenseA010ImagePublisherQueue, that keeps the current
implementation in which images are published to listeners as in the
producer-consumer design pattern.
